+++ src/funcx_common/redis/__init__.py
@@ -1,0 +1,7 @@
+from .connection import FuncxRedisConnection
+from .task_queue import FuncxEndpointTaskQueue
+
+__all__ = (
+    "FuncxRedisConnection",
+    "FuncxEndpointTaskQueue",
+)