+++ src/funcx_common/tasks/__init__.py
@@ -1,0 +1,7 @@
+from .constants import TaskState
+from .protocol import TaskProtocol
+
+__all__ = (
+    "TaskState",
+    "TaskProtocol",
+)