+++ src/funcx_common/tasks/constants.py
@@ -1,0 +1,12 @@
+import enum
+
+
+# subclass from str so that the enum can be JSON-encoded without adjustment
+class TaskState(str, enum.Enum):
+    RECEIVED = "received"  # on receiving a task web-side
+    WAITING_FOR_EP = "waiting-for-ep"  # while waiting for ep to accept/be online
+    WAITING_FOR_NODES = "waiting-for-nodes"  # jobs are pending at the scheduler
+    WAITING_FOR_LAUNCH = "waiting-for-launch"
+    RUNNING = "running"
+    SUCCESS = "success"
+    FAILED = "failed"